:truck: Flatten documenten component code
