+++ src/openzaak/components/besluiten/tests/test_audittrails.py
@@ -7,7 +7,7 @@ from vng_api_common.tests import reverse
 from vng_api_common.utils import get_uuid_from_path
 
 from openzaak.components.catalogi.models.tests.factories import BesluitTypeFactory
-from openzaak.components.documenten.models.tests.factories import (
+from openzaak.components.documenten.tests.factories import (
     EnkelvoudigInformatieObjectFactory,
 )
 from openzaak.utils.tests import JWTAuthMixin
