+++ src/openzaak/components/besluiten/tests/test_besluitinformatieobjecten.py
@@ -3,7 +3,7 @@ from rest_framework.test import APITestCase
 from vng_api_common.tests import get_validation_errors, reverse, reverse_lazy
 from vng_api_common.validators import IsImmutableValidator
 
-from openzaak.components.documenten.models.tests.factories import (
+from openzaak.components.documenten.tests.factories import (
     EnkelvoudigInformatieObjectFactory,
 )
 from openzaak.utils.tests import JWTAuthMixin
