+++ src/openzaak/components/documenten/api/kanalen.py
@@ -1,8 +1,6 @@
-from django.conf import settings
-
 from vng_api_common.notifications.kanalen import Kanaal
 
-from openzaak.components.documenten.models import EnkelvoudigInformatieObject
+from ..models import EnkelvoudigInformatieObject
 
 KANAAL_DOCUMENTEN = Kanaal(
     "documenten",
