+++ src/openzaak/components/documenten/tests/models/test_statussen.py
@@ -6,8 +6,8 @@ from datetime import date
 from django.core.exceptions import ValidationError
 from django.test import TestCase
 
-from ..constants import Statussen
-from .factories import EnkelvoudigInformatieObjectFactory
+from ...constants import Statussen
+from ..factories import EnkelvoudigInformatieObjectFactory
 
 
 class StatusTests(TestCase):
