+++ src/openzaak/components/documenten/tests/test_eio_file.py
@@ -16,13 +16,14 @@ from vng_api_common.tests import reverse
 from openzaak.components.catalogi.models.tests.factories import (
     InformatieObjectTypeFactory,
 )
-from openzaak.components.documenten.api.tests.utils import get_operation_url
-from openzaak.components.documenten.models import EnkelvoudigInformatieObject
-from openzaak.components.documenten.models.tests.factories import (
+from openzaak.utils.tests import JWTAuthMixin
+
+from ..models import EnkelvoudigInformatieObject
+from .factories import (
     EnkelvoudigInformatieObjectCanonicalFactory,
     EnkelvoudigInformatieObjectFactory,
 )
-from openzaak.utils.tests import JWTAuthMixin
+from .utils import get_operation_url
 
 
 @override_settings(SENDFILE_BACKEND="sendfile.backends.simple")
