+++ src/openzaak/components/documenten/tests/test_gebruiksrechten.py
@@ -2,12 +2,13 @@ from rest_framework import status
 from rest_framework.test import APITestCase
 from vng_api_common.tests import get_validation_errors, reverse
 
-from openzaak.components.documenten.models.tests.factories import (
+from openzaak.utils.tests import JWTAuthMixin
+
+from .factories import (
     EnkelvoudigInformatieObjectCanonicalFactory,
     EnkelvoudigInformatieObjectFactory,
     GebruiksrechtenFactory,
 )
-from openzaak.utils.tests import JWTAuthMixin
 
 
 class GebruiksrechtenTests(JWTAuthMixin, APITestCase):
