+++ src/openzaak/components/documenten/tests/test_notifications_send.py
@@ -12,8 +12,9 @@ from vng_api_common.tests import reverse
 from openzaak.components.catalogi.models.tests.factories import (
     InformatieObjectTypeFactory,
 )
-from openzaak.components.documenten.api.tests.utils import get_operation_url
 from openzaak.utils.tests import JWTAuthMixin
+
+from .utils import get_operation_url
 
 
 @freeze_time("2012-01-14")
