+++ src/openzaak/components/zaken/api/tests/test_validation.py
@@ -27,7 +27,7 @@ from openzaak.components.catalogi.models.tests.factories import (
     ZaakInformatieobjectTypeFactory,
     ZaakTypeFactory,
 )
-from openzaak.components.documenten.models.tests.factories import (
+from openzaak.components.documenten.tests.factories import (
     EnkelvoudigInformatieObjectFactory,
 )
 from openzaak.components.zaken.models import ZaakInformatieObject
