+++ src/openzaak/components/zaken/models/tests/factories.py
@@ -9,7 +9,6 @@ from vng_api_common.constants import (
     ZaakobjectTypes,
 )
 
-from openzaak.components.besluiten.tests.factories import BesluitFactory
 from openzaak.components.catalogi.models.tests.factories import (
     EigenschapFactory,
     ResultaatTypeFactory,
@@ -17,7 +16,7 @@ from openzaak.components.catalogi.models.tests.factories import (
     StatusTypeFactory,
     ZaakTypeFactory,
 )
-from openzaak.components.documenten.models.tests.factories import (
+from openzaak.components.documenten.tests.factories import (
     EnkelvoudigInformatieObjectCanonicalFactory,
 )
